# docs/swagger.yaml
basePath: /
definitions:
...
  api.PlayoutStatus:
    properties:
      aqueue:
        format: uint64
        type: integer
      debug: {}
      drop:
        format: uint64
        type: integer
      dup:
        format: uint64
        type: integer
      duplicating:
        type: boolean
      enc:
        format: uint64
        type: integer
      gop:
        type: string
      id:
        type: string
      input:
        $ref: '#/definitions/api.PlayoutStatusIO'
      looping:
        type: boolean
      output:
        $ref: '#/definitions/api.PlayoutStatusIO'
      queue:
        format: uint64
        type: integer
      stream:
        format: uint64
        type: integer
      swap:
        $ref: '#/definitions/api.PlayoutStatusSwap'
      url:
        type: string
    type: object
  api.PlayoutStatusIO:
    properties:
      packet:
        format: uint64
        type: integer
      size_kb:
        format: uint64
        type: integer
      state:
        enum:
        - running
        - idle
        type: string
      time:
        format: uint64
        type: integer
    type: object
  api.PlayoutStatusSwap:
    properties:
      lasterror:
        type: string
      lasturl:
        type: string
      status:
        type: string
      url:
        type: string
    type: object
...
  errors.Error:
    properties:
      code:
        format: int
        type: integer
      details:
        items:
          type: string
        type: array
      message:
        type: string
    type: object
...
info:
  contact:
    email: hello@datarhei.com
    name: datarhei Core Support
    url: https://www.datarhei.com
  description: Expose REST API for the datarhei Core
  license:
    name: Apache 2.0
    url: https://github.com/datarhei/core/v16/blob/main/LICENSE
  title: datarhei Core API
  version: "3.0"
paths:
...
  /api/v3/process/{id}/playout/{inputid}/errorframe/{name}:
    post:
      consumes:
      - application/octet-stream
      description: Upload an error frame which will be encoded immediately
      operationId: process-3-playout-errorframe
      parameters:
      - description: Process ID
        in: path
        name: id
        required: true
        type: string
      - description: Process Input ID
        in: path
        name: inputid
        required: true
        type: string
      - description: Any filename with a suitable extension
        in: path
        name: name
        required: true
        type: string
      - description: Image to be used a error frame
        in: body
        name: image
        required: true
        schema:
          items:
            type: integer
          type: array
      produces:
      - text/plain
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - ApiKeyAuth: []
      summary: Upload an error frame
      tags:
      - v16.7.2
  /api/v3/process/{id}/playout/{inputid}/errorframe/encode:
    get:
      description: Immediately encode the errorframe (if available and looping)
      operationId: process-3-playout-errorframencode
      parameters:
      - description: Process ID
        in: path
        name: id
        required: true
        type: string
      - description: Process Input ID
        in: path
        name: inputid
        required: true
        type: string
      produces:
      - text/plain
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - ApiKeyAuth: []
      summary: Encode the errorframe
      tags:
      - v16.7.2
  /api/v3/process/{id}/playout/{inputid}/keyframe/{name}:
    get:
      description: Get the last keyframe of an input of a process. The extension of
        the name determines the return type.
      operationId: process-3-playout-keyframe
      parameters:
      - description: Process ID
        in: path
        name: id
        required: true
        type: string
      - description: Process Input ID
        in: path
        name: inputid
        required: true
        type: string
      - description: Any filename with an extension of .jpg or .png
        in: path
        name: name
        required: true
        type: string
      produces:
      - image/jpeg
      - image/png
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: file
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - ApiKeyAuth: []
      summary: Get the last keyframe
      tags:
      - v16.7.2
  /api/v3/process/{id}/playout/{inputid}/reopen:
    get:
      description: Close the current input stream such that it will be automatically
        re-opened
      operationId: process-3-playout-reopen-input
      parameters:
      - description: Process ID
        in: path
        name: id
        required: true
        type: string
      - description: Process Input ID
        in: path
        name: inputid
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - ApiKeyAuth: []
      summary: Close the current input stream
      tags:
      - v16.7.2
  /api/v3/process/{id}/playout/{inputid}/status:
    get:
      description: Get the current playout status of an input of a process
      operationId: process-3-playout-status
      parameters:
      - description: Process ID
        in: path
        name: id
        required: true
        type: string
      - description: Process Input ID
        in: path
        name: inputid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.PlayoutStatus'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - ApiKeyAuth: []
      summary: Get the current playout status
      tags:
      - v16.7.2
  /api/v3/process/{id}/playout/{inputid}/stream:
    put:
      consumes:
      - text/plain
      description: Replace the current stream with the one from the given URL. The
        switch will only happen if the stream parameters match.
      operationId: process-3-playout-stream
      parameters:
      - description: Process ID
        in: path
        name: id
        required: true
        type: string
      - description: Process Input ID
        in: path
        name: inputid
        required: true
        type: string
      - description: URL of the new stream
        in: body
        name: url
        required: true
        schema:
          type: string
      produces:
      - text/plain
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - ApiKeyAuth: []
      summary: Switch to a new stream
      tags:
      - v16.7.2
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  ApiRefreshKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  Auth0KeyAuth:
    in: header
    name: Authorization
    type: apiKey
  BasicAuth:
    type: basic
swagger: "2.0"